Another Fix for path_mapping config not loading
Run the functions that access the config in the main thread by using the
self.timeout method built into the session.

I know there is already another solution to this problem, but I was
curious if making the access to the configuration run on the main thread
had any benefits over storing the path_mapping beforehand.